+++ src/main/java/com/github/correa/finalreality/model/character/player/CharacterClass.java
@@ -4,10 +4,10 @@ package com.github.correa.finalreality.model.character.player;
  * Enumeration of the classes a player character may have.
  *
  * @author Ignacio Slater Muñoz.
- * @author <Your name>
+ * @author <Benjamín Correa>
  */
 public enum CharacterClass {
-  KNIGHT, ENGINEER, THIEF, BLACK_MAGE, WHITE_MAGE, ENEMY
+  KNIGHT, ENGINEER, THIEF, BLACK_MAGE, WHITE_MAGE
 }